refactor: modularize puzzle and add routing
