+++ src/pages/Home.jsx
@@ -1,0 +1,9 @@
+import Puzzle from "../puzzle/Puzzle.jsx";
+
+export default function Home() {
+  return (
+    <div className="min-h-screen bg-slate-950 text-slate-100">
+      <Puzzle />
+    </div>
+  );
+}
