+++ src/puzzle/rng.js
@@ -1,0 +1,18 @@
+function mulberry32(seed) {
+  let t = seed >>> 0;
+  return function () {
+    t += 0x6D2B79F5;
+    let r = Math.imul(t ^ (t >>> 15), 1 | t);
+    r ^= r + Math.imul(r ^ (r >>> 7), 61 | r);
+    return ((r ^ (r >>> 14)) >>> 0) / 4294967296;
+  };
+}
+
+export function makeRng(seed) {
+  const rand = mulberry32(seed);
+  return {
+    next: () => rand(),
+    int: (max) => Math.floor(rand() * max),
+    choice: (arr) => arr[Math.floor(rand() * arr.length)],
+  };
+}
